requirements.txt
fastapi
httpx
langchain
langchain-community
langchain-experimental
langchain-openai
lxml
markdown
numpy
pandas
pillow
pydantic
pytesseract
requests
tabula-py
uvicorn
python-dotenv
cssselect
